# internal/web/static/openapi.yaml
openapi: 3.0.3
info:
  title: Knomit API
  version: 2.0.0
  description: |
    Git-backed knowledge base API (HAL/HATEOAS).

    All successful responses are `application/hal+json` (except SSE streams
    and the raw OpenAPI spec). Collection responses use an envelope with
    `count`, `_links`, and `_embedded.<rel>`. All error responses are
    RFC 9457 `application/problem+json`.
servers:
  - url: /api/v1
    description: v1
...
paths:
...
  /repos/{repo}/branches/{branch}/facts/{path}:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/Branch"
      - $ref: "#/components/parameters/FactPath"
    get:
      tags: [facts]
      summary: Read a fact (anchored at HEAD)
      description: |
        When the path terminates with `/incoming` or `/outgoing`, the response
        is a `RefsCollection` of referencing/referenced facts instead of a
        fact view. Each entry carries a `commit` field pinning it to a
        specific version (source for `/incoming`, target for `/outgoing`)
        and `_links.self` is anchored to that commit.
      operationId: getFact
      responses:
        "200":
          description: Fact or ref collection
          content:
            application/hal+json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/Fact" }
                  - { $ref: "#/components/schemas/RefsCollection" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
    put:
      tags: [facts]
      summary: Update a fact (full content replace)
      operationId: updateFact
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FactUpdateRequest" }
      responses:
        "200":
          description: Updated fact
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/Fact" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [facts]
      summary: Delete a fact
      operationId: deleteFact
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
  /repos/{repo}/branches/{branch}/facts/{path}/commits:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/Branch"
      - $ref: "#/components/parameters/FactPath"
    get:
      tags: [facts, commits]
      summary: Per-fact commit log
      description: |
        Returns commits that touched the given fact path, in reverse chronological
        order. Used by the version walker (right-pane fact header) and the commit
        drawer's "this fact · N versions" section.
      operationId: factCommits
      parameters:
        - { name: limit,  in: query, description: "1–500; out-of-range or non-integer is 400.", schema: { type: integer, default: 50, minimum: 1, maximum: 500 } }
        - { name: after,  in: query, schema: { type: string }, description: cursor }
        - { name: from,   in: query, schema: { type: string } }
        - { name: before, in: query, schema: { type: string } }
      responses:
        "200":
          description: Collection of commits touching the fact
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/CommitsCollection" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
...
  /repos/{repo}/branches/{branch}/domains:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/Branch"
    get:
      tags: [stats]
      summary: Domain histogram
      operationId: listDomains
      responses:
        "200":
          description: Collection of domains
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/DomainsCollection" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
  /repos/{repo}/branches/{branch}/domains/{name}:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/Branch"
      - { name: name, in: path, required: true, schema: { type: string } }
    get:
      tags: [stats, facts]
      summary: Facts by domain
      operationId: getDomainFacts
      responses:
        "200":
          description: Facts in the domain
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/FactsCollection" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
...
  /repos/{repo}/branches/{branch}/commits:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/Branch"
    get:
      tags: [commits]
      summary: List commits
      operationId: listCommits
      parameters:
        - { name: limit,  in: query, description: "1–500; out-of-range or non-integer is 400.", schema: { type: integer, default: 50, minimum: 1, maximum: 500 } }
        - { name: after,  in: query, schema: { type: string }, description: cursor }
        - { name: from,   in: query, schema: { type: string } }
        - { name: before, in: query, schema: { type: string } }
      responses:
        "200":
          description: Collection of commits
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/CommitsCollection" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
  /repos/{repo}/branches/{branch}/commits/{sha}:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/Branch"
      - $ref: "#/components/parameters/Sha"
    get:
      tags: [commits]
      summary: Commit detail with affected files
      operationId: getCommit
      responses:
        "200":
          description: Commit
          content:
            application/hal+json:
              schema: { $ref: "#/components/schemas/CommitDetail" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
...
  /repos/{repo}/origin-sessions/{sessionID}/commit:
    parameters:
      - $ref: "#/components/parameters/Repo"
      - $ref: "#/components/parameters/SessionID"
    post:
      tags: [origin, events]
      summary: Finalize connection (SSE stream)
      operationId: commitOriginSession
      responses:
        "200":
          description: SSE stream (phases swapping → configuring → rebuilding → done)
          content:
            text/event-stream:
              schema: { type: string }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
components:
  parameters:
    Repo:       { name: repo,      in: path, required: true, schema: { type: string } }
    Branch:     { name: branch,    in: path, required: true, schema: { type: string } }
    Sha:        { name: sha,       in: path, required: true, schema: { type: string } }
...
    SessionID:  { name: sessionID, in: path, required: true, schema: { type: string } }
...
    FactPath:
      name: path
      in: path
      required: true
      description: Fact path under the ontology root (may contain slashes).
      schema: { type: string }
  responses:
    Problem:
      description: RFC 9457 problem+json
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Link:
      type: object
      required: [href]
      properties:
        href:      { type: string }
        templated: { type: boolean }
    LinkMap:
      type: object
      additionalProperties: { $ref: "#/components/schemas/Link" }
    CollectionBase:
      type: object
      required: [count, _links]
      properties:
        count:     { type: integer }
        _links:    { $ref: "#/components/schemas/LinkMap" }
        _embedded: { type: object }
    Problem:
      type: object
      required: [title, status]
      properties:
        type:     { type: string, default: "about:blank" }
        title:    { type: string }
        status:   { type: integer }
        detail:   { type: string }
        instance: { type: string }

    # Discovery / repos / branches
...
    FactAsOf:
      type: object
      properties:
        branch: { type: string }
        commit: { type: string }
    Fact:
      type: object
      required: [path, title, _links]
      properties:
        path:       { type: string }
        title:      { type: string }
        body:       { type: string }
        type:
          type: string
          enum: [observation, concept, process, principle, pattern, reference, synthesis, insight, hypothesis, methodology]
        domain:     { type: array, items: { type: string } }
        entities:   { type: array, items: { type: string } }
        refs:
          type: array
          items: { $ref: "#/components/schemas/Ref" }
        confidence: { type: number, format: double }
        sources:   { type: integer }
        as_of:     { $ref: "#/components/schemas/FactAsOf" }
        _links:    { $ref: "#/components/schemas/LinkMap" }
    FactSummary:
      type: object
      required: [path, title, _links]
      properties:
        path:         { type: string }
        title:        { type: string }
        type:         { type: string }
        confidence:   { type: number, format: double }
        committed_at: { type: string, format: date-time }
        operation:    { type: string }
        _links:       { $ref: "#/components/schemas/LinkMap" }
    Ref:
      type: object
      required: [path]
      properties:
        path:    { type: string }
        title:   { type: string }
        type:    { type: string }
        # commit pins each entry to a specific version: source_commit for
        # /incoming items (the commit that asserted the ref), target_commit
        # for /outgoing items (the version of the target the ref resolved to
        # at the source's commit). _links.self is anchored to this commit.
        commit:  { type: string, description: "Commit hash this entry is pinned to (source_commit on /incoming, target_commit on /outgoing)." }
        # Outgoing-only: target Fact node is soft-deleted at the anchor.
        # When true, _links.self is omitted (no navigable destination).
        deleted: { type: boolean }
        _links:  { $ref: "#/components/schemas/LinkMap" }
    FactsCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionBase"
        - type: object
          properties:
            _embedded:
              type: object
              properties:
                facts:
                  type: array
                  items: { $ref: "#/components/schemas/FactSummary" }
    RefsCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionBase"
        - type: object
          properties:
            _embedded:
              type: object
              properties:
                refs:
                  type: array
                  items: { $ref: "#/components/schemas/Ref" }
...
    FactUpdateRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: Full markdown document with YAML frontmatter.

    # Topics
...
    DomainEntry:
      type: object
      required: [name, count, _links]
      properties:
        name:   { type: string }
        count:  { type: integer }
        _links: { $ref: "#/components/schemas/LinkMap" }
    DomainsCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionBase"
        - type: object
          properties:
            _embedded:
              type: object
              properties:
                domains:
                  type: array
                  items: { $ref: "#/components/schemas/DomainEntry" }
...
    FileCounts:
      type: object
      properties:
        added:    { type: integer }
        modified: { type: integer }
        deleted:  { type: integer }
    CommitAuthor:
      type: object
      description: >
        The git author identity of the commit, recorded verbatim. For agent
        writes, name is the agent-id and email carries the +operation
        subaddress (e.g. agent-7+learn@agents.knomit.io). For PR merges into
        main, this is the human who merged (the committer would be GitHub).
      required: [name, email]
      properties:
        name:  { type: string }
        email: { type: string }
    CommitSummary:
      type: object
      required: [commit, date, author, _links]
      properties:
        commit:    { type: string }
        date:      { type: string, format: date-time }
        message:   { type: string }
        operation: { type: string }
        author:    { $ref: "#/components/schemas/CommitAuthor" }
        files:     { $ref: "#/components/schemas/FileCounts" }
        _links:    { $ref: "#/components/schemas/LinkMap" }
    CommitsCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionBase"
        - type: object
          properties:
            _embedded:
              type: object
              properties:
                commits:
                  type: array
                  items: { $ref: "#/components/schemas/CommitSummary" }
    CommitFile:
      type: object
      required: [path, action]
      properties:
        path:   { type: string }
        action: { type: string, enum: [added, modified, deleted] }
        title:  { type: string }
        _links: { $ref: "#/components/schemas/LinkMap" }
    CommitDetail:
      type: object
      required: [commit, date, author, _links]
      properties:
        commit:    { type: string }
        date:      { type: string, format: date-time }
        message:   { type: string }
        operation: { type: string }
        author:    { $ref: "#/components/schemas/CommitAuthor" }
        files:
          type: array
          items: { $ref: "#/components/schemas/CommitFile" }
        _links: { $ref: "#/components/schemas/LinkMap" }

    # Jobs
...